+++ CHANGELOG.md
@@ -2374,3 +2374,4 @@
 - [2026-06-26 16:48] perf(test): apply RepaintBoundary on animated test widgets
 - [2026-06-26 16:59] fix(analytics): prevent double tap duplicate action in analytics
 - [2026-06-26 17:13] feat(theme): implement responsive theme view component
+- [2026-06-26 17:18] feat(ui): implement shimmer loading skeleton in ui